refactor(plugin-js-packages): split pkg mng helper for testing
