+++ packages/plugin-js-packages/src/lib/constants.ts
@@ -1,9 +1,5 @@
 import { IssueSeverity } from '@code-pushup/models';
-import type {
-  DependencyGroup,
-  PackageAuditLevel,
-  PackageManagerId,
-} from './config';
+import type { DependencyGroup, PackageAuditLevel } from './config';
 import { DependencyGroupLong } from './runner/outdated/types';
 
 export const defaultAuditLevelMapping: Record<
